@@ -1,3 +1,4 @@
+using BaGet.Core.Server.Transformers;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.AspNetCore.Routing.Constraints;
@@ -6,40 +7,25 @@ namespace BaGet.Extensions
 {
     public static class IRouteBuilderExtensions
     {
-        public static IRouteBuilder MapServiceIndexRoutes(this IRouteBuilder routes)
+        public static IEndpointRouteBuilder MapServiceIndexRoutes(this IEndpointRouteBuilder endpoints)
         {
-            return routes.MapRoute(
-                name: Routes.IndexRouteName,
-                template: "v3/index.json",
-                defaults: new { controller = "ServiceIndex", action = "GetAsync" });
+            endpoints.MapDynamicControllerRoute<NugetServiceIndexRoutesTransformer>("v3/index.json");
+            return endpoints;
         }
 
-        public static IRouteBuilder MapPackagePublishRoutes(this IRouteBuilder routes)
+        public static IEndpointRouteBuilder MapPackagePublishRoutes(this IEndpointRouteBuilder endpoints)
         {
-            routes.MapRoute(
-                name: Routes.UploadPackageRouteName,
-                template: "api/v2/package",
-                defaults: new { controller = "PackagePublish", action = "Upload" },
-                constraints: new { httpMethod = new HttpMethodRouteConstraint("PUT") });
-
-            routes.MapRoute(
-                name: Routes.DeleteRouteName,
-                template: "api/v2/package/{id}/{version}",
-                defaults: new { controller = "PackagePublish", action = "Delete" },
-                constraints: new { httpMethod = new HttpMethodRouteConstraint("DELETE") });
-
-            routes.MapRoute(
-                name: Routes.RelistRouteName,
-                template: "api/v2/package/{id}/{version}",
-                defaults: new { controller = "PackagePublish", action = "Relist" },
-                constraints: new { httpMethod = new HttpMethodRouteConstraint("POST") });
-
-            return routes;
+            endpoints.MapDynamicControllerRoute<NugetPackagePublishRoutesTransformer>("api/v2/package/{id?}/{version?}");
+            return endpoints;
         }
 
-        public static IRouteBuilder MapSymbolRoutes(this IRouteBuilder routes)
+        public static IEndpointRouteBuilder MapSymbolRoutes(this IEndpointRouteBuilder endpoints)
         {
-            routes.MapRoute(
+            endpoints.MapDynamicControllerRoute<NugetSymbolRoutesTransformer>("api/v2/symbol");
+            endpoints.MapDynamicControllerRoute<NugetSymbolRoutesTransformer>("api/v2/symbol/{first}/{second}/{thrid}/{forth?}");
+            endpoints.MapDynamicControllerRoute<NugetSymbolRoutesTransformer>("api/v2/symbols/{first}/{second}/{thrid}/{forth?}");
+            return endpoints;
+            /*routes.MapRoute(
                 name: Routes.UploadSymbolRouteName,
                 template: "api/v2/symbol",
                 defaults: new { controller = "Symbol", action = "Upload" },
@@ -51,36 +37,25 @@ namespace BaGet.Extensions
                 defaults: new { controller = "Symbol", action = "Get" });
             
             routes.MapRoute(
-                name: Routes.SymbolDownloadRouteName,
+                name: Routes.SymbolDownloadRouteNamePrefix,
                 template: "api/download/symbols/{prefix}/{file}/{key}/{file2}",
                 defaults: new { controller = "Symbol", action = "Get" });
 
-            return routes;
+            return routes;*/
         }
 
-        public static IRouteBuilder MapSearchRoutes(this IRouteBuilder routes)
+        public static IEndpointRouteBuilder MapSearchRoutes(this IEndpointRouteBuilder endpoints)
         {
-            routes.MapRoute(
-                name: Routes.SearchRouteName,
-                template: "v3/search",
-                defaults: new { controller = "Search", action = "SearchAsync" });
-
-            routes.MapRoute(
-                name: Routes.AutocompleteRouteName,
-                template: "v3/autocomplete",
-                defaults: new { controller = "Search", action = "AutocompleteAsync" });
-
-            // This is an unofficial API to find packages that depend on a given package.
-            routes.MapRoute(
-                name: Routes.DependentsRouteName,
-                template: "v3/dependents",
-                defaults: new { controller = "Search", action = "DependentsAsync" });
-
-            return routes;
+            endpoints.MapDynamicControllerRoute<NugetSearchRoutesTransformer>("v3/{action}");
+            return endpoints;
         }
 
-        public static IRouteBuilder MapPackageMetadataRoutes(this IRouteBuilder routes)
+        public static IEndpointRouteBuilder MapPackageMetadataRoutes(this IEndpointRouteBuilder endpoints)
         {
+            endpoints.MapDynamicControllerRoute<NugetPackageMetadataRoutesTransformer>("v3/registration/{id}/index.json");
+            endpoints.MapDynamicControllerRoute<NugetPackageMetadataRoutesTransformer>("v3/registration/{id}/{version}.json");
+            return endpoints;
+            /*
             routes.MapRoute(
                name: Routes.RegistrationIndexRouteName,
                template: "v3/registration/{id}/index.json",
@@ -91,11 +66,17 @@ namespace BaGet.Extensions
                 template: "v3/registration/{id}/{version}.json",
                 defaults: new { controller = "PackageMetadata", action = "RegistrationLeafAsync" });
 
-            return routes;
+            return routes;*/
         }
 
-        public static IRouteBuilder MapPackageContentRoutes(this IRouteBuilder routes)
+        public static IEndpointRouteBuilder MapPackageContentRoutes(this IEndpointRouteBuilder endpoints)
         {
+            endpoints.MapDynamicControllerRoute<NugetPackageContentRoutesTransformer>("v3/package/{id}/{ext}");
+            endpoints.MapDynamicControllerRoute<NugetPackageContentRoutesTransformer>("v3/package/{id}/{version}/{idVersion}.nupkg");
+            endpoints.MapDynamicControllerRoute<NugetPackageContentRoutesTransformer>("v3/package/{id}/{version}/{id2}.nuspec");
+            endpoints.MapDynamicControllerRoute<NugetPackageContentRoutesTransformer>("v3/package/{id}/{version}/readme");
+            return endpoints;
+            /*
             routes.MapRoute(
                 name: Routes.PackageVersionsRouteName,
                 template: "v3/package/{id}/index.json",
@@ -116,7 +97,7 @@ namespace BaGet.Extensions
                 template: "v3/package/{id}/{version}/readme",
                 defaults: new { controller = "PackageContent", action = "DownloadReadmeAsync" });
 
-            return routes;
+            return routes;*/
         }
     }
 }
