@@ -3,9 +3,11 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Cors.Infrastructure;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ApplicationParts;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
+using System.Reflection;
 
 namespace BaGet.Core.Server.Extensions
 {
@@ -13,11 +15,15 @@ namespace BaGet.Core.Server.Extensions
     {
         public static IServiceCollection ConfigureHttpServices(this IServiceCollection services)
         {
+            var controllersAssembly = Assembly.GetAssembly(typeof(BaGet.Controllers.ServiceIndexController));
+            var part = new AssemblyPart(controllersAssembly);
             services
-                .AddMvc()
-                .AddApplicationPart(typeof(BaGet.Controllers.PackageContentController).Assembly)
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2)
-                .AddJsonOptions(options =>
+                .AddControllers()
+                .ConfigureApplicationPartManager(apm => apm.ApplicationParts.Add(part))
+                .AddApplicationPart(controllersAssembly)
+                .AddControllersAsServices()
+                .SetCompatibilityVersion(CompatibilityVersion.Version_3_0)
+                .AddNewtonsoftJson(options =>
                 {
                     options.SerializerSettings.DateTimeZoneHandling = DateTimeZoneHandling.Utc;
                 });
