@@ -6,11 +6,17 @@ using Microsoft.Extensions.Logging;
 
 namespace BaGet
 {
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+        }
+    }
     /// <summary>
     /// Captures <see cref="OperationCanceledException" /> and converts to HTTP 409 response.
     /// Based off: https://github.com/aspnet/AspNetCore/blob/28157e62597bf0e043bc7e937e44c5ec81946b83/src/Middleware/Diagnostics/src/DeveloperExceptionPage/DeveloperExceptionPageMiddleware.cs
     /// </summary>
-    public class OperationCancelledMiddleware
+        public class OperationCancelledMiddleware
     {
         private readonly RequestDelegate _next;
         private readonly ILogger<OperationCancelledMiddleware> _logger;
